Bot behaviour.
Further developed bot behaviour; they now don't crash the game
upon death. Also, the PRBot ticker has been moved to a proper
location, to avoid tmthing-related crashes.

Several states are now complete, but the fields 'avoid' and
'want' are still completely unused. Not read to, not written to
(except when resetting a bot).