Clean the PrBot to a mostly playable state
The PrBot code was integrated into the rest of the codebase, which did
involve minor edits to g_game.c (to exempt local bots from consistency
checks) and m_cheat.c (to provide the player with a command to add them
ingame).

Many fixes were put into place, which rectified various scenarios that
used to result in hard errors, as well as gave the bot a modicum of
liveliness by wandering and shooting at enemies.

At this point, it should be possible to watch a bot shoot at any
enemies it sees, which is already not a lot since it doesn't really
always wander (even when it is doing nothing else), and never tends to
make actual progress into the level that way either.

I want to consolidate the bot's movement and behaviour, before moving
on to less tedious things, like pathing and actually doing interesting
things. Until then... sigh.